fix(nats): harden sub-agent tool registration
Start configured sub-agent toolsets concurrently so the registration barrier is bounded by the slowest server instead of cumulative startup latency.

Centralize reserved sub-agent session tool names, require nextest for environment-sensitive coverage, and preserve NATS worker test cohesion. This addresses the CodeRabbit and CodeScene findings from #1492.
